backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.12
anthropic==0.40.0
lark-oapi==1.3.5
python-dotenv==1.0.1
httpx==0.27.2
pycryptodome==3.21.0
PyYAML>=6.0
APScheduler>=3.10
